requirements.txt
Package             Version
------------------- -------
asgiref             3.8.1
backports.zoneinfo  0.2.1
beautifulsoup4      4.13.3
crispy-bootstrap5   2025.4
Django              4.2.20
django-bootstrap-v5 1.0.11
django-crispy-forms 2.3
pillow              10.4.0
pip                 25.0.1
setuptools          75.3.2
soupsieve           2.6
sqlparse            0.5.3
typing_extensions   4.13.1
tzdata              2025.2
